Added pdo_test.php to test making PDO connections to MySQL
In pdo_test, only the currently active line
```
$stmt = $dbh->query('SELECT * FROM Apprentices');
```
succesfully selects the correct data from the db, and returns the array of results.

The other pairs of ```$stmt = $dbh->prepare(``` and ```$stmt->bindValue(``` statements,
as well as the ```$dbh->query``` statement that uses the $tableName variable, all attempt
to pass the explicit SQL string contained in the single quotes.